chore: move media stuff from documentActions/documentsActions to a new mediaActions
